fixed DateHelper YEAR_MS, needs to be long instead of int
git-svn-id: https://141.76.40.86/svn-students/iircommon/toolkit/trunk@1159 5654db93-19e5-4567-beb8-64e8c25c7c45

Former-commit-id: 97a783dc523dd7a6ec254aefec78aef5498e3a28

+++ src/main/java/tud/iir/extraction/ExtractionProcessManager.java
@@ -248,8 +248,8 @@ public class ExtractionProcessManager {
      * @param totalTimeMS The total pause time in milliseconds.
      * @param intervalMS The interval in milliseconds when the loop should write live status updates.
      */
-    private static void waitingLoop(int totalTimeMS, int intervalMS) {
-        int steps = totalTimeMS / intervalMS;
+    private static void waitingLoop(long totalTimeMS, long intervalMS) {
+        long steps = totalTimeMS / intervalMS;
         for (int i = 0; i < steps; i++) {
             ThreadHelper.sleep(intervalMS);
             int progress = (int) Math.floor((double) 100 * i / steps);
